dev/pyproject.toml
[project]
name = "react-agent"
version = "0.1.0"
description = "Deployment of LangChain Playground"
authors = [
    { name = "Teron", email = "t3ron.ay@gmail.com" },
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.11,<3.12"
dependencies = [
    "docling",
    "fal-client",
    "google-genai",
    "httpx",
    "ipython",
    "langchain",
    "langchain-community",
    "langchain-google-genai",
    "langchain-openai",
    "langgraph",
    "langgraph-prebuilt",
    "langsmith",
    "more-itertools",
    "numpy",
    "opencc-python-reimplemented",
    "pandas",
    "pydantic",
    "pydub",
    "python-dotenv",
    "pytubefix",
    "replicate",
    "rich",
    "tavily-python",
    "tqdm",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]
# Explicitly exclude openai-whisper or specify a compatible version
# to avoid the dependency on numba which doesn't support Python 3.12
whisper = ["openai-whisper<20240930"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph.templates.react_agent", "react_agent", "Tools"]

[tool.setuptools.package-dir]
"langgraph.templates.react_agent" = "src/react_agent"
"react_agent" = "src/react_agent"
"Tools" = "src/Tools"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "langgraph-cli[inmem]",
]
